Set up Astro View Transitions
To enable swish image transitions between featured posts and post pages,
I needed to move post images to the frontmatter. I wanted to avoid doing
this, because I want to be able to break the format sometimes. I tried
creating an astro component for post hero images, but the challenge
there was uniquely identifying them in the featured posts and post
markup (the slug seemed hard to get in the MDX scope). Rendering the
post title / image via frontmatter was the simplest way.